syncmem => index
